+++ src/compy/index.js
@@ -1,6 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+export * from '#x/BaseAggregateComponent';
 export * from '#x/BaseComponent';
 export * from '#x/BaseConfig';
 export * from '#x/ControlContext';