+++ src/jest.config.mjs
@@ -16,7 +16,9 @@ export default {
     collectCoverage: false,
     collectCoverageFrom: [
       '**/code/node_modules/**',
-      '!**/*.test.*'
+      '!**/*.test.*',
+      '!**/Intf[A-Z]*.{js,cjs,mjs}',
+      '!**/Type[A-Z]*.{js,cjs,mjs}'
     ],
     coverageDirectory: `${OUT_DIR}/coverage`,
     coveragePathIgnorePatterns: [],